+++ src/scenes/LevelDuckyScene.ts
@@ -44,6 +44,7 @@ export class LevelDuckyScene extends BaseLevelScene {
       mentorDialog: MENTOR_DIALOG.ducky,
       startingPlayerTexture: "adventurer-bubbles",
       bgMusic: [330, 392, 494, 587, 494, 392],
+      hideRewardCaption: true,
     };
   }
 }